backend/requirement.txt
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
Authlib==1.3.2
babel==2.16.0
bcrypt==4.2.1
billiard==4.2.1
blinker==1.9.0
boto3==1.35.77
botocore==1.35.77
Brotli==1.1.0
cachelib==0.13.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.0
cssselect2==0.7.0
Deprecated==1.2.15
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
filelock==3.16.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.9.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Moment==1.0.6
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.55.2
google-auth==2.36.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
limits==3.14.1
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgspec==0.18.6
mysql-connector-python==9.1.0
numpy==2.2.0
oauthlib==3.2.2
opencv-python==4.10.0.84
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydyf==0.11.0
Pygments==2.18.0
PyJWT==2.10.1
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
s3transfer==0.10.4
six==1.17.0
slugify==0.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
vendor==0.1
vine==5.1.0
wcwidth==0.2.13
weasyprint==63.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.0
WTForms==3.2.1
zipp==3.21.0
zopfli==0.2.3.post1
